GUI: support for single notification template
