--- dojo/db_migrations/0146_enable_notification_templates.py
@@ -1,34 +1,0 @@
-# Generated by Django 3.2.11 on 2022-01-19 15:28
-
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('dojo', '0145_system_settings_default_group_email_pattern'),
-    ]
-
-    operations = [
-        # migrations.RemoveConstraint(
-        #     model_name='notifications',
-        #     name='notifications_user_product',
-        # ),
-        # migrations.RemoveIndex(
-        #     model_name='notifications',
-        #     name='dojo_notifi_user_id_575989_idx',
-        # ),
-        migrations.AddField(
-            model_name='notifications',
-            name='template',
-            field=models.CharField(default=None, max_length=100, null=True),
-        ),
-        migrations.AddIndex(
-            model_name='notifications',
-            index=models.Index(fields=['user', 'product', 'template'], name='dojo_notifi_user_id_994807_idx'),
-        ),
-#        migrations.AddConstraint(
-#            model_name='notifications',
-#            constraint=models.UniqueConstraint(fields=('user', 'product', 'template'), name='notifications_user_product_template'),
-#        ),
-    ]
